chore(set): move to `lib`.
