+++ src/set/lib/index.ts
@@ -1,0 +1,6 @@
+
+export { CoreSet } from './core-set.abstract';
+export { DataSet } from './data-set.class';
+export { DataWeakSet } from '../../weak-set/data-weak-set.class';
+export { ImmutableSet } from './immutable-set.class';
+export { ImmutableWeakSet } from '../../weak-set/immutable-weak-set.class';
